Commit section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageZize = 5;
   loading = false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService : MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -43,8 +44,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(
-      ()=> this.messages.splice(this.messages.findIndex(m=> m.id ==id),1));
+    this.confirmService.confirm('Confirm Delete message','Sure you want to delete?').subscribe(result=>
+      {
+        if(result)
+        {
+          this.messageService.deleteMessage(id).subscribe(
+            ()=> this.messages.splice(this.messages.findIndex(m=> m.id ==id),1));
+        }
+      })
+   
   }
 
 }
